+++ apps/commudle-admin/src/app/feature-modules/policies/policies.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { PoliciesRoutingModule } from './policies-routing.module';
 import { NbCardModule } from '@commudle/theme';
-import { PoliciesComponent } from './policies/policies.component';
+import { PoliciesComponent } from './component/policies/policies.component';
 
 @NgModule({
   declarations: [PoliciesComponent],
